requirements.txt
openai
python-dotenv
pinecone
pytelegrambotapi